C2 - modularização inicial (#2)
* build: :ambulance: separação em componentes do estado inicial do projeto

* fix: :art: tava a main toda junto
